+++ openedx_events/testing.py
@@ -1,5 +1,7 @@
 """
-Utils used by Open edX event tests.
+Test utilities for use by consumers of openedx-events.
+
+Provides mixins that help isolate Open edX event state between test classes.
 """
 
 from openedx_events.tooling import OpenEdxPublicSignal
@@ -100,11 +102,11 @@ class OpenEdxEventsTestMixin(EventsIsolationMixin):
 
     Example usage:
 
-        class MyTestCase(TestCase, OpenEdxEventsTestCase):
+        class MyTestCase(TestCase, OpenEdxEventsTestMixin):
 
             ENABLED_OPENEDX_EVENTS = ['org.openedx.learning.student.registration.completed.v1']
 
-    This class assumes that's it's being used in conjunction TestCase or TestCase subclasses.
+    This class assumes that it's being used in conjunction with TestCase or TestCase subclasses.
     """
 
     ENABLED_OPENEDX_EVENTS = []
