+++ openedx_events/tests/test_generate_avro_schemas.py
@@ -8,7 +8,7 @@ from django.test import TestCase
 
 from openedx_events.event_bus.avro.tests.test_utilities import create_simple_signal
 from openedx_events.management.commands.generate_avro_schemas import Command
-from openedx_events.tests.utils import FreezeSignalCacheMixin
+from openedx_events.testing import FreezeSignalCacheMixin
 from openedx_events.tooling import KNOWN_UNSERIALIZABLE_SIGNALS, OpenEdxPublicSignal, load_all_signals
